requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues


azure-functions
mysql.connector
datetime
# uuid
pandas
flask
twilio
azure.storage.blob
azure.identity
azure.keyvault.secrets
schedule
python-dotenv
werkzeug
openpyxl
requests